final commit

+++ database/migrations/2025_02_15_114301_create_folders_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('name')->nullable();
             $table->string('path')->nullable();
             $table->index('path');
-            $table->integer('size')->nullable();
+            $table->bigInteger('size')->nullable();
             $table->timestamps();
         });
     }
